core-handler/requirements.txt
redis
sqlalchemy
python-dotenv
psycopg2-binary
python-jose
